Implemented second (broad) component.
Might still need some polish (it definitely does), but it works.
One problem is that the fitter seems to have trouble converging on
good values when there are too many parameters - we might simply need
a more capable (and thus slow, but...) minimizer, and none such is
implemented in GModelFit yet.