Now stores 0th moment as well as fitted fluxes
Each "Slice" is now a N x M x 5 array rather than being 3 deep as
before. The mid one is a little dumb and might be either used or removed
later on, so that may introduce breaking changes later on, but I think
we are still at a point where that is permissible.

So now, results for each line are saved in five slices:
1. Fitted line flux
2. Fitted line flux std. dev.
3. Fitted line flux S/N (really necessary though?
4. 0th moment
5. 0th moment std. dev.