Implemented first steps of Voronoi binning
First step: A wrapper that allows to easily voronoi bin a flux (or
other) slice from a slices_dict.

Also wrote a quick function to take the average (including errors and
NaNs) in a given fragmentation map, e.g. a Voronoi bin, based on an
image with one or more numbered masks.

I have not yet found a good way to cycle through bins and fit them.
I have tried, and it kinda-sorta-works, but it is *extremely* slow.
So let's call this a "preview", shall we?